fix: modular plugins
